% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Description of the Dataset \code{posterior.rda}}
\format{A data frame of class "bhrcr".}
\usage{
data("posterior")
}
\description{
The output given by \code{clearanceEstimatorBayes} in the slow demo example, which is calculated
and saved ahead of time to save users' time.
}
\keyword{datasets}
