% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pursat.R
\name{pursat}
\alias{pursat}
\title{Description of the Dataset \code{pursat.rda}}
\format{A data frame with following variables
\describe{
  \item{id}{patients' IDs}
  \item{time}{measurement times}
  \item{count}{parasite counts in blood}
}}
\usage{
data("pursat")
}
\description{
The \code{pursat} dataset consists of P. falciparum
clearance profiles of 110 patients, measured in 2009 and 
2010 in the Pursat province of Western Cambodia. Parasite densities were measured every 6 hours, 
and the detection limit was 15 parasites per microliter.
All 110 individuals were observed until no parasites were found in their blood.
}
\keyword{datasets}
